Adds win condition
Adds a timer to keep time once happiness levels reach 80 or higher, and once it passes 3 consecutive minutes, it will make a pop up appear that alerts the user that they have won. Also adds the timer function to the _updateHappiness function so once happiness updates to 80 it would start tracking.